+++ pkg/autoscaler/aggregation/bucketing_test.go
@@ -75,32 +75,15 @@ func TestTimedFloat64BucketsSimple(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			buckets := NewTimedFloat64Buckets(tt.granularity)
+			// New implementation test.
+			buckets := NewTimedFloat64Buckets(2*time.Minute, tt.granularity)
 			for _, stat := range tt.stats {
 				buckets.Record(stat.time, stat.name, stat.value)
 			}
 
 			got := make(map[time.Time]float64)
-			for time, bucket := range buckets.buckets {
-				got[time] = bucket
-			}
-
-			if !cmp.Equal(tt.want, got) {
-				t.Errorf("Unexpected values (-want +got): %v", cmp.Diff(tt.want, got))
-			}
-			if len(tt.want) == 0 && !buckets.isEmpty() {
-				t.Error("IsEmpty() = false, want true")
-			}
-
-			// New implementation test.
-			buckets2 := NewTimedFloat64Buckets2(2*time.Minute, tt.granularity)
-			for _, stat := range tt.stats {
-				buckets2.Record(stat.time, stat.name, stat.value)
-			}
-
-			got = make(map[time.Time]float64)
 			// Less time in future than our window is (2mins above), but more than any of the tests report.
-			buckets2.ForEachBucket(trunc1.Add(time.Minute), func(t time.Time, b float64) {
+			buckets.ForEachBucket(trunc1.Add(time.Minute), func(t time.Time, b float64) {
 				// Since we're storing 0s when there's no data, we need to exclude those
 				// for this test.
 				if b > 0 {
@@ -111,9 +94,6 @@ func TestTimedFloat64BucketsSimple(t *testing.T) {
 			if !cmp.Equal(tt.want, got) {
 				t.Errorf("Unexpected values (-want +got): %v", cmp.Diff(tt.want, got))
 			}
-			if len(tt.want) == 0 && !buckets.isEmpty() {
-				t.Error("IsEmpty() = false, want true")
-			}
 		})
 	}
 }
@@ -121,13 +101,11 @@ func TestTimedFloat64BucketsSimple(t *testing.T) {
 func TestTimedFloat64BucketsManyReps(t *testing.T) {
 	granularity := time.Second
 	trunc1 := time.Now().Truncate(granularity)
-	buckets := NewTimedFloat64Buckets(granularity)
-	buckets2 := NewTimedFloat64Buckets2(time.Minute, granularity)
+	buckets := NewTimedFloat64Buckets(time.Minute, granularity)
 	for p := 0; p < 5; p++ {
 		trunc1 = trunc1.Add(granularity)
 		for t := 0; t < 5; t++ {
 			buckets.Record(trunc1, pod+strconv.Itoa(p), float64(p+t))
-			buckets2.Record(trunc1, pod+strconv.Itoa(p), float64(p+t))
 		}
 	}
 	// So the buckets are:
@@ -141,7 +119,7 @@ func TestTimedFloat64BucketsManyReps(t *testing.T) {
 	buckets.ForEachBucket(trunc1, func(_ time.Time, b float64) {
 		sum1 += b
 	})
-	buckets2.ForEachBucket(trunc1, func(_ time.Time, b float64) {
+	buckets.ForEachBucket(trunc1, func(_ time.Time, b float64) {
 		sum2 += b
 	})
 	if got, want := sum1, want; got != want {
@@ -156,14 +134,10 @@ func TestTimedFloat64BucketsManyReps(t *testing.T) {
 func TestTimedFloat64BucketsForEachBucket(t *testing.T) {
 	granularity := time.Second
 	trunc1 := time.Now().Truncate(granularity)
-	buckets := NewTimedFloat64Buckets(granularity)
-	buckets2 := NewTimedFloat64Buckets2(2*time.Minute, granularity)
+	buckets := NewTimedFloat64Buckets(2*time.Minute, granularity)
 
-	if buckets.ForEachBucket(trunc1, func(time time.Time, bucket float64) {}) {
-		t.Fatalf("ForEachBucket unexpectedly returned non-empty result")
-	}
 	// Since we recorded 0 data, even in this implementation no iteration must occur.
-	if buckets2.ForEachBucket(trunc1, func(time time.Time, bucket float64) {}) {
+	if buckets.ForEachBucket(trunc1, func(time time.Time, bucket float64) {}) {
 		t.Fatalf("ForEachBucket unexpectedly returned non-empty result")
 	}
 
@@ -172,36 +146,10 @@ func TestTimedFloat64BucketsForEachBucket(t *testing.T) {
 	buckets.Record(trunc1.Add(2*time.Second), pod, 5.0)
 	buckets.Record(trunc1.Add(3*time.Second), pod, 5.0)
 
-	buckets2.Record(trunc1, pod, 10.0)
-	buckets2.Record(trunc1.Add(1*time.Second), pod, 10.0)
-	buckets2.Record(trunc1.Add(2*time.Second), pod, 5.0)
-	buckets2.Record(trunc1.Add(3*time.Second), pod, 5.0)
-
 	acc1 := 0
 	acc2 := 0
-	if !buckets.ForEachBucket(trunc1,
-		func(time.Time, float64) {
-			acc1++
-		},
-		func(time.Time, float64) {
-			acc2++
-		},
-	) {
-		t.Fatal("ForEachBucket unexpectedly returned empty result")
-	}
 
-	want := 4
-	if acc1 != want {
-		t.Errorf("acc1 = %v, want %v", acc1, want)
-	}
-	if acc2 != want {
-		t.Errorf("acc2 = %v, want %v", acc1, want)
-	}
-
-	// Now verify second impl
-	acc1 = 0
-	acc2 = 0
-	if !buckets2.ForEachBucket(trunc1.Add(4*time.Second),
+	if !buckets.ForEachBucket(trunc1.Add(4*time.Second),
 		func(_ time.Time, b float64) {
 			// We need to exclude the 0s for this test.
 			if b > 0 {
@@ -216,102 +164,12 @@ func TestTimedFloat64BucketsForEachBucket(t *testing.T) {
 	) {
 		t.Fatal("ForEachBucket unexpectedly returned empty result")
 	}
-
-	if acc1 != want {
-		t.Errorf("acc1 = %v, want %v", acc1, want)
-	}
-	if acc2 != want {
-		t.Errorf("acc2 = %v, want %v", acc1, want)
-	}
-}
-
-func TestTimedFloat64BucketsRemoveOlderThan(t *testing.T) {
-	pod := "pod"
-	zero := time.Now()
-	trunc1 := zero.Truncate(1 * time.Second)
-
-	tests := []struct {
-		name            string
-		granularity     time.Duration
-		times           []time.Time
-		removeOlderThan time.Time
-		want            []time.Time
-	}{{
-		name:        "remove one",
-		granularity: 1 * time.Second,
-		times: []time.Time{
-			trunc1,
-			trunc1.Add(1 * time.Second),
-			trunc1.Add(2 * time.Second),
-		},
-		removeOlderThan: trunc1.Add(1 * time.Second),
-		want: []time.Time{
-			trunc1.Add(1 * time.Second),
-			trunc1.Add(2 * time.Second),
-		},
-	}, {
-		name:        "remove all",
-		granularity: 1 * time.Second,
-		times: []time.Time{
-			trunc1,
-			trunc1.Add(1 * time.Second),
-		},
-		removeOlderThan: trunc1.Add(2 * time.Second),
-		want:            []time.Time{},
-	}, {
-		name:        "remove none",
-		granularity: 1 * time.Second,
-		times: []time.Time{
-			trunc1,
-			trunc1.Add(1 * time.Second),
-		},
-		removeOlderThan: trunc1,
-		want: []time.Time{
-			trunc1,
-			trunc1.Add(1 * time.Second),
-		},
-	}, {
-		name:            "empty",
-		granularity:     1 * time.Second,
-		times:           []time.Time{},
-		removeOlderThan: trunc1.Add(1 * time.Second),
-		want:            []time.Time{},
-	}}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			buckets := NewTimedFloat64Buckets(tt.granularity)
-			for _, time := range tt.times {
-				buckets.Record(time, pod, 1.0)
-			}
-
-			if got, want := len(buckets.buckets), len(tt.times); got != want {
-				t.Errorf("len(buckets) = %v, want %v", got, want)
-			}
-
-			buckets.RemoveOlderThan(tt.removeOlderThan)
-
-			if got, want := len(buckets.buckets), len(tt.want); got != want {
-				t.Errorf("len(buckets) = %v, want %v", got, want)
-			}
-
-			got := make(map[time.Time]bool)
-			for time := range buckets.buckets {
-				got[time] = true
-			}
-			for _, want := range tt.want {
-				if !got[want] {
-					t.Errorf("Expected buckets to contain %v, buckets: %v", want, got)
-				}
-			}
-		})
-	}
 }
 
 func TestTimedFloat64BucketsWindowUpdate(t *testing.T) {
 	granularity := time.Second
 	trunc1 := time.Now().Truncate(granularity)
-	buckets := NewTimedFloat64Buckets2(5*time.Second, granularity)
+	buckets := NewTimedFloat64Buckets(5*time.Second, granularity)
 
 	// Fill the whole bucketing list with rollover.
 	buckets.Record(trunc1, pod, 1)
@@ -382,7 +240,7 @@ func TestTimedFloat64BucketsWindowUpdate3sGranularity(t *testing.T) {
 	trunc1 := time.Now().Truncate(granularity)
 
 	// So two buckets here (ceil(5/3)=ceil(1.6(6))=2).
-	buckets := NewTimedFloat64Buckets2(5*time.Second, granularity)
+	buckets := NewTimedFloat64Buckets(5*time.Second, granularity)
 	if got, want := len(buckets.buckets), 2; got != want {
 		t.Fatalf("Initial bucket count = %d, want: %d", got, want)
 	}
